use 8 bytes hash instead of 4 byte hash, because of hash collisions, if many files are used (>100000)

+++ src/mcachefs-types.h
@@ -38,7 +38,8 @@ typedef unsigned long long int hash_t;
 #endif
 
 #ifdef __MCACHEFS_HASH_USE_32_6_16
-typedef unsigned int hash_t;
+//typedef unsigned int hash_t;
+typedef unsigned long long int hash_t;
 #define __MCACHEFS_HASH_ALGORITHM "32.6.16"
 #endif
